Refactor image helpers
