+++ simaas/plugins/builtins/rti_aws/service.py
@@ -10,7 +10,7 @@ from typing import Optional, Tuple, Dict, List
 import boto3
 from pydantic import BaseModel
 
-from simaas.cli.cmd_image import clone_repository, build_processor_image
+from simaas.core.image import clone_repository, build_processor_image
 from simaas.core.helpers import get_timestamp_now
 
 from simaas.core.logging import get_logger
